requirements.txt
Flask==3.0.3

transformers~=4.50.3
torch~=2.6.0
pillow~=11.1.0
ctransformers~=0.2.27